fix: Boden-Kollision in MoveIt, Not-Aus-Verriegelung, Nexus-Zugriffsschutz
Cartesian Jogging / Boden
- neuer Node moveit_floor_collision: Tischebene als Kollisionsobjekt auf
  /planning_scene (move_group + servo_server), in fake/realmove-Launch
  eingebunden, abschaltbar mit floor_collision:=false
- Robot Control UI bremst Jogging weich bis zur Z Collision Level (15 mm)
  ab, weil MoveIt Servo (Humble) nach einer Kollision in alle Richtungen
  fast stehen bleibt; Gelenk-Jog sperrt die Richtung zum Tisch
- manuelle MoveTo-Ziele unter 15 mm werden abgelehnt

Sicherheit
- Jog-Buttons stoppen bei pointerleave/-cancel, blur, Tab-Wechsel und
  rosbridge-Abbruch (Z/Rotation liefen vorher weiter)
- Not-Aus bleibt verriegelt bis /ui/reset_emergency_stop, UI zeigt dafuer
  einen Quittieren-Button; Servo ruht bis zum Quittieren
- Motion Handler: atomare Ausfuehrungssperre, Autostart ueber die Sperre
  (auto_initial_pose), gueltige Stopp-Trajektorie ohne Gelenkdaten
- Gamepad-Checker sperrt "nach unten" ohne aktuelle EEF-Position
- Nexus Webapp: POST nur lokal und nur von der eigenen Origin, CORS nur
  fuer GET, Terminal-Titel ohne Shell-String, Zombie-Prozesse eingesammelt
- certs/ aus dem Repo genommen, VR-Launch erzeugt fehlende Zertifikate

Aufraeumen / Optimierung
- TF-Listener in den Grasp Executors nur noch einmal pro Node
- Dashboard: Executable-Cache mit einem Aufruf, ungenutzte Videos und
  Altdateien entfernt, sauberes Beenden des Analyzers
- verwaiste Eintraege in launcher_config.json entfernt
- install_zed.sh: cuda-keyring statt apt-key
- Compiler-Warnungen in xarm_moveit_servo und zed_components behoben

+++ src/zed-ros2-wrapper/zed_components/src/zed_camera/src/zed_camera_component.cpp
@@ -4135,8 +4135,8 @@ bool ZedCamera::startCamera()
   if (mSvoMode) {
     RCLCPP_INFO(
       get_logger(), " * SVO resolution\t-> %ldx%ld",
-      mZed->getCameraInformation().camera_configuration.resolution.width,
-      mZed->getCameraInformation().camera_configuration.resolution.height);
+      static_cast<long>(mZed->getCameraInformation().camera_configuration.resolution.width),
+      static_cast<long>(mZed->getCameraInformation().camera_configuration.resolution.height));
     RCLCPP_INFO_STREAM(
       get_logger(),
       " * SVO framerate\t-> "
